Use new IBM Watson voices
* The old demo website and voices were removed from IBM’s website
* Most exisiting voices now only have a “V3” available.
* A number of new languages and voices are now available.
* TODO: Pitch and speed control

NOTE: We can no longer grab the resulting audio file with a GET request. Instead we need to use proxy.php and send a POST request, which results in us receiving raw audio (mp3) data. For convenience this is base64 encoded so that the <audio> tag can play and we don’t need to save a file locally.